Extract version command
