+++ pkg/cmd/version/version_test.go
@@ -1,4 +1,4 @@
-package root
+package version
 
 import (
 	"testing"
